Update Exercise 3 Q1.cpp

+++ Exercise 3 Q1.cpp
@@ -6,10 +6,13 @@ int main()
 	float price = 9.90 ;
 	int quantity ;
 	const int DISC = 10 ;
-	float totalprice;
+	double totalprice;
 	cout << "The quantity of disc purchased : ";
 	cin  >> quantity ;
+	
+	//calculation
 	totalprice = (price * quantity) * (100 - DISC) / 100.00;
+	
 	cout << "The total price is " << totalprice << '\n';
 	system("pause");
 	return 0;
